+++ src/__init__.py
@@ -2,5 +2,5 @@
 metadata for the package
 """
 
-__version__ = "1.7.3"
+__version__ = "1.8.0"
 __author__ = "Edmundo Garcia: ingega@gmail.com"