+++ src/api/v1/apps/orders/schemas.py
@@ -23,6 +23,10 @@ class FlavorCatalogueCreate(FlavorCatalogueBase):
 class FlavorCatalogueRead(FlavorCatalogueBase):
     id: int
 
+class FlavorCatalogueUpdate(SQLModel):
+    product_id: Optional[int] = None
+    description: Optional[str] = Field(default=None, max_length=50)
+
 
 # ==========================================
 # 2. PRODUCT SCHEMAS
@@ -44,7 +48,6 @@ class ProductUpdate(SQLModel):
     main_dish: Optional[str] = Field(default=None, max_length=50)
     category: Optional[str] = Field(default=None, max_length=30)
     price: Optional[condecimal(max_digits=6, decimal_places=2)] = Field(default=None)
-
 
 
 class ProductWithFlavors(ProductRead):